A Step Towards Movement
So movement has appeared within the game. It's very ugly right now and
will be fixed up with awesome animations, sprites, visual effects, and
more. The backbone of what I think will be 90% of the classes in the
game have been built because movement required it. Much more information
can be seen in my journal, which I don't have the heart to copy and
paste here. hmmm... some ideas and questions were tossed around in my
head and written down for my meeting with everybody tomorrow.