Another Step Toward Movement
After a bit of testing and making sure all interactions were happening
properly, I have gotten more of movement finished, and one more day
should see all of it (what is currently in the game) complete. At the
moment, clicking the triage area will move the nurse there, and then
open information about a patient. Then the player can choose to send a
patient to an exam room, waiting room, and simply kick them out. When I
work on this next, I need to add some UI for the exam room, and make
sure the buttons work for both the triage and waiting chairs.