New Classes
Added some new classes Diagnosis and ABG to the game. ABG is a class
that will perform the majority of functions related to ABG values.
Diagnosis is the class that each patient will reference. Both of these
will simplify and de-clutter existing classes, while meeting
final-product goals of allowing the game to be edited without the use of
Unity.