Assessments
I built the interface for the new Exam Room Computer. I implemented the
functionality spoken about and updated several scripts and methods in
the process so that they will be ready for more changes that are soon to
come. The new interface has been added into the game and everything that
I had the capability to give functionality to is there. More in Journal