XML Integration - First Steps
Worked on my previous XML document, and made a change suggested by
professor Warren. Added the functionality for the ABG class to read in
the different nursing interventions from xml and create the diagnoses
from the input data. Diagnosis class has been updated for the new format
of data being pulled from the xml file, and the examroomUI had an error
fixed that was resulting in a null reference error on the new diagnosis
class. Found a formatting problem with the layout elements, but was
unable to solve it.