Update for July 21st Demo
Updated the XML with the translations Warren had access to. Reformatted
the xml for Nursing Interventions. Rewrote the XML Reading in the ABG
script for the new format of the Nursing Intervention XML. Added in the
Translated versions of the ABG Answers (RM, AA, C). Rewrote some of the
functionality of the Diagnosis class to work with the new xml. Diagnoses
now have greetings for each language that are displayed in the waiting
room. Added Direct Translation to the LanguageManager script for
specific texts, such as ABG Answers (RM, AA, C). Added in the Dropdown
Button Translations.  Reference Desk practice tool now displays the
correct terms for the chosen language. Game Over screen is now
functional and in the game. Patient Feedback screen is now functional
and in the game. Did some testing to determine whether or not I was
deep/shallow cloning. It's shallow, so I'll have to fix that. Added a
button to the main menu to allow players to change the current language.
Added a back button to the reference desk/practice scene. It's stricly
placeholder, and its hard to miss. Tussled with the Webplayer Build for
quite a while to make sure it was reading in the XML files/data
properly.